1.21 port

+++ common/src/main/java/org/figuramc/figura/mixin/render/CameraMixin.java
@@ -26,7 +26,7 @@ public abstract class CameraMixin {
     @Unique private Avatar avatar;
 
     @Shadow protected abstract void setRotation(float yaw, float pitch);
-    @Shadow protected abstract void move(double x, double y, double z);
+    @Shadow protected abstract void move(float x, float y, float z);
 
     @Inject(method = "setup", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Camera;setRotation(FF)V", shift = At.Shift.AFTER))
     private void setupRot(BlockGetter area, Entity focusedEntity, boolean thirdPerson, boolean inverseView, float tickDelta, CallbackInfo ci) {
@@ -117,7 +117,7 @@ public abstract class CameraMixin {
         if (avatar != null) {
             FiguraVec3 pos = avatar.luaRuntime.renderer.cameraPos;
             if (pos != null && pos.notNaN())
-                move(-pos.z, pos.y, -pos.x);
+                move((float) -pos.z, (float) pos.y, (float) -pos.x);
 
             avatar = null;
         }
